+++ flatlaf-core/src/main/java/com/formdev/flatlaf/util/SystemInfo.java
@@ -56,6 +56,7 @@ public class SystemInfo
 
 	// other
 	/** @since 1.1 */ public static final boolean isProjector;
+	/** @since 1.1.1 */ public static final boolean isWinPE;
 
 	static {
 		// platforms
@@ -91,6 +92,7 @@ public class SystemInfo
 
 		// other
 		isProjector = Boolean.getBoolean( "org.jetbrains.projector.server.enable" );
+		isWinPE = isWindows && "X:\\Windows\\System32".equalsIgnoreCase( System.getProperty( "user.dir" ) );
 	}
 
 	public static long scanVersion( String version ) {